+++ Classes-IPhone/GHUnitIPhoneAppDelegate.m
@@ -11,36 +11,15 @@
 
 @implementation GHUnitIPhoneAppDelegate
 
-- (void)runTests {	
-	[NSThread detachNewThreadSelector:@selector(_runTests) toTarget:self withObject:nil];	
-}
-
-- (void)_runTests {
-	NSAutoreleasePool *pool = [[NSAutoreleasePool alloc] init];
-	GHTestRunner *runner = [[[GHTestRunner runnerFromEnv] retain] autorelease];
-	runner.delegate = self;
-	// To allow exceptions to raise into the debugger, uncomment below
-	//runner.raiseExceptions = YES;
-	
-	[runner run];
-	[pool release];
-}
-
 - (void)applicationDidFinishLaunching:(UIApplication *)application {
-	viewController_ = [[GHUnitIPhoneViewController alloc] init];
+	viewController_ = [[GHUnitIPhoneViewController alloc] init];	
+	
 	navigationController_ = [[UINavigationController alloc] initWithRootViewController:viewController_];
 	CGSize size = [[UIScreen mainScreen] applicationFrame].size;
 	window_ = [[UIWindow alloc] initWithFrame:CGRectMake(0, 0, size.width, size.height)];
 	GHUDebug(@"Setting window view");
 	[window_ addSubview:navigationController_.view];
 	[window_ makeKeyAndVisible];
-	
-	// Defaults
-	[[NSUserDefaults standardUserDefaults] registerDefaults:
-	 [NSDictionary dictionaryWithObject:[NSNumber numberWithBool:YES] forKey:@"GHUnit-Autorun"]];
-	
-	if ([[[NSUserDefaults standardUserDefaults] objectForKey:@"GHUnit-Autorun"] boolValue]) 	
-		[self runTests];
 }
 
 - (void)dealloc {
@@ -48,33 +27,6 @@
 	[viewController_ release];
 	[window_ release];
 	[super dealloc];
-}
-
-#pragma mark Delegates (GHTestRunner)
-
-- (void)testRunner:(GHTestRunner *)runner didLog:(NSString *)message {
-	[viewController_ setStatusText:message];
-}
-
-- (void)testRunner:(GHTestRunner *)runner test:(id<GHTest>)test didLog:(NSString *)message {
-
-}
-
-- (void)testRunner:(GHTestRunner *)runner didStartTest:(id<GHTest>)test {
-	[viewController_ updateTest:test];
-}
-
-- (void)testRunner:(GHTestRunner *)runner didFinishTest:(id<GHTest>)test {
-	[viewController_ updateTest:test];
-}
-
-- (void)testRunnerDidStart:(GHTestRunner *)runner { 
-	[viewController_ setGroup:(id<GHTestGroup>)runner.test];
-}
-
-- (void)testRunnerDidFinish:(GHTestRunner *)runner {
-	GHTestStats stats = [runner.test stats];
-	[viewController_ setTestStats:stats];
 }
 
 @end